A very slow crawler

+++ src/MongoDBAdapter.java
@@ -14,17 +14,33 @@ import com.mongodb.client.MongoCollection;
 import com.mongodb.client.MongoCursor;
 import com.mongodb.client.MongoDatabase;
 import com.mongodb.client.model.Filters;
+import org.apache.commons.lang3.StringUtils;
 
 import org.bson.*;
 
 public class MongoDBAdapter {
-	public MongoDBAdapter() {}
+	
 	private MongoCollection<Document> UnvisitedCollection;
 	private MongoCollection<Document> VisitedCollection;
+	private MongoCollection<Document> RobotCollection;
 	private MongoClientURI uri = new MongoClientURI( "mongodb+srv://CrawlerDB:Crawler123@crawlercluster-kgwg6.mongodb.net/Test?retryWrites=true&w=majority"  ); ;
-	private MongoClient mongoClient = new MongoClient(uri);
-	private MongoDatabase database = mongoClient.getDatabase("Test");
-	public void init() {
+//	private MongoClient mongoClient = new MongoClient(uri);
+//	private MongoDatabase database = mongoClient.getDatabase("Test");
+//	private MongoClient mongoClient = new MongoClient( "localhost" , 27017 );
+//	private MongoDatabase database = mongoClient.getDatabase("Crawler");
+	private MongoClient mongoClient;
+	private MongoDatabase database;
+	
+	public MongoDBAdapter(boolean Global) {
+		if(Global) {
+			this.mongoClient = new MongoClient(uri);
+			this.database = mongoClient.getDatabase("Test");
+		} else {
+			this.mongoClient = new MongoClient( "localhost" , 27017 );
+			this.database = mongoClient.getDatabase("Crawler");
+		}
+	}
+	public void init(boolean DropTables) {
 		Logger.getLogger("org.mongodb.driver").setLevel(Level.WARNING);
 		
 			  
@@ -32,40 +48,49 @@ public class MongoDBAdapter {
 			  
 		      System.out.println("Database Connected Successfully ...");
 		      
-		      // Deleting Old Collections if Exists
-		      MongoCollection<Document> OLDUnvisited = database.getCollection("Unvisited");
-		      OLDUnvisited.drop();
-		      MongoCollection<Document> OLDVisited = database.getCollection("Visited");
-		      OLDVisited.drop();
+		      if(DropTables) {
+			      // Deleting Old Collections if Exists
+			      MongoCollection<Document> OLDUnvisited = database.getCollection("Unvisited");
+			      OLDUnvisited.drop();
+			      MongoCollection<Document> OLDVisited = database.getCollection("Visited");
+			      OLDVisited.drop();
+			      MongoCollection<Document> OLDRobot = database.getCollection("Robot");
+			      OLDRobot.drop();
+			      
+			      // Creating a new Collections
+			      database.createCollection("Unvisited");
+			      database.createCollection("Visited");
+			      database.createCollection("Robot");
+		      }
 		      
-		      // Creating a new Unvisited and Visited Collections
-		      database.createCollection("Unvisited");
-		      database.createCollection("Visited");
 		      UnvisitedCollection = database.getCollection("Unvisited");
 		      VisitedCollection = database.getCollection("Visited");
+		      RobotCollection = database.getCollection("Robot");
 		      
-		      //Reading Initial Seed Set from File
-		      List<Document> documents = new ArrayList<Document>();
-			  BufferedReader reader;
-			  Document docTemp;
-			  System.out.println("Initiallizing Seed Set with ...");
-			  try {
-					reader = new BufferedReader(new FileReader((".\\Sports Sites.txt")));
-					String line = reader.readLine();
-					while (line != null) {
-						
-						docTemp = new Document("url", line);
-						System.out.println("Adding "+line+" to Unvisited");
-						documents.add(docTemp);
-						// read next line
-							line = reader.readLine();
-						}
-					reader.close();
-			} catch (IOException e) {
-				e.printStackTrace();
-			}
-		    // Insert Multiple Documents
-			  UnvisitedCollection.insertMany(documents);
+		      if (DropTables) {
+			      //Reading Initial Seed Set from File
+			      List<Document> documents = new ArrayList<Document>();
+				  BufferedReader reader;
+				  Document docTemp;
+				  System.out.println("Initiallizing Seed Set with ...");
+				  try {
+						reader = new BufferedReader(new FileReader((".\\Sports Sites.txt")));
+						String line = reader.readLine();
+						while (line != null) {
+							line = line.replaceAll("[^:]//", "/").toLowerCase();
+							docTemp = new Document("url", line);
+							System.out.println("Adding "+line+" to Unvisited");
+							documents.add(docTemp);
+							// read next line
+								line = reader.readLine();
+							}
+						reader.close();
+				} catch (IOException e) {
+					e.printStackTrace();
+				}
+			    // Insert Multiple Documents
+				  UnvisitedCollection.insertMany(documents);
+		      }
 			  
 		  }
 				
@@ -83,7 +108,7 @@ public class MongoDBAdapter {
 				url = result.get("url").toString();	
 			}
 			UnvisitedCollection.deleteOne(Filters.eq("url",url));
-			System.out.println("Removing "+url+" from Visited");
+			System.out.println("Removing "+url+" from Unvisited");
 			System.out.println("Remaining unvisited urls: "+UnvisitedCollection.countDocuments());
 			return url;
 		}
@@ -102,10 +127,12 @@ public class MongoDBAdapter {
 		List<Document> documents = new ArrayList<Document>();
 		Document docTemp;
 		for(String URI : URIs){
+			URI = URI.replaceAll("[^:]//", "/").toLowerCase();
 			Document result1 = UnvisitedCollection.find(Filters.eq("url",URI)).first();
 			Document result2 = VisitedCollection.find(Filters.eq("url",URI)).first();
 			
 			if(result1 == null &&  result2 == null) {
+
 				docTemp = new Document("url", URI);
 				System.out.println("Adding "+URI+" to Unvisited");
 				documents.add(docTemp);
@@ -119,10 +146,13 @@ public class MongoDBAdapter {
 	
 	public boolean addUnvisited(String URI) {
 		
+		URI = URI.replaceAll("[^:]//", "/").toLowerCase();
 		Document result1 = UnvisitedCollection.find(Filters.eq("url",URI)).first();
 		Document result2 = VisitedCollection.find(Filters.eq("url",URI)).first();
 		
 		if(result1 == null &&  result2 == null) {
+			
+			URI.toLowerCase();
 			Document docTemp = new Document("url", URI);
 			System.out.println("Adding "+URI+" to Unvisited");
 			UnvisitedCollection.insertOne(docTemp);
@@ -131,6 +161,66 @@ public class MongoDBAdapter {
 		return false;
 	}
 	
+	public int addRobots(List<String> URIs) {
+		List<Document> documents = new ArrayList<Document>();
+		Document docTemp;
+		for(String URI : URIs){
+			URI = URI.replaceAll("[^:]//", "/").toLowerCase();
+			Document result = RobotCollection.find(Filters.eq("urlRegex",URI)).first();
+			
+			if(result == null) {
+				docTemp = new Document("urlRegex", URI);
+				System.out.println("Adding "+URI+" to Robot");
+				documents.add(docTemp);
+			} else {
+				continue;
+			}
+		}
+		RobotCollection.insertMany(documents);
+		return documents.size();
+	}
+	
+	public boolean inRobots(String URI)	{
+		FindIterable<Document> result = RobotCollection.find();
+		URI = URI.replaceAll("[^:]//", "/").toLowerCase();
+		System.out.println("Check this in Robot "+ URI);
+		for(Document currRegex : result)
+		{
+			String regex = currRegex.get("urlRegex").toString();
+			
+			if(URI.matches(regex)) {
+				return true;
+			}
+		}		
+		return false;
+	}
+	
+	public long visitedCount() {return VisitedCollection.countDocuments();}
+	
+	
+	@SuppressWarnings("deprecation")
+	public boolean addVisited(String URI , String content) {
+		FindIterable<Document> result = VisitedCollection.find();
+		System.out.println("here11");
+		for(Document url : result)
+		{
+			String content2 = url.get("Document").toString().toLowerCase();
+			String url2 = url.get("url").toString();
+			
+			int distance = StringUtils.getLevenshteinDistance(content.toLowerCase(), content2);
+			float percentage = (float)distance * 100 / content.length(); 
+			
+			if(percentage < 20 || url2.equalsIgnoreCase(URI)) {
+				System.out.println("Content Matched to " + url2);
+				return false;
+			}				
+		}
+		System.out.println("here12");
+		Document docTemp = new Document("url", URI).append("Document", content);
+		VisitedCollection.insertOne(docTemp);
+		System.out.println("here13");
+		return true;
+	}
 
 	public static void main( String args[] ) {
